Add some tests and re-organize
